prefer the avrdude shipped by arduino sdk
Since cmake prefers the systems paths instead of the given ones, the
avrdude installed by some Linux distributions might be prefered. Since
arduino patches the avrdude program, the default should be to use the
avrdude provided by the arduino sdk instead of the system one and only
use the later as a fallback option.

Of course, the user can still override the ARDUINO_AVRDUDE_PROGRAM
variable within cmake to use the systems binary.